refactor: move into methods module
